# Export: better performance by avoiding pointless image->python data->image roundtrips