--- a/addons/io_scene_gltf2/blender/exp/gltf2_blender_image.py
+++ b/addons/io_scene_gltf2/blender/exp/gltf2_blender_image.py
@@ -14,121 +14,199 @@
 
 import bpy
 import os
-import typing
+from typing import Optional
 import numpy as np
 import tempfile
+import enum
+
+
+class Channel(enum.IntEnum):
+    R = 0
+    G = 1
+    B = 2
+    A = 3
+
+# These describe how an ExportImage's channels should be filled.
+
+class FillImage:
+    """Fills a channel with the channel src_chan from a Blender image."""
+    def __init__(self, image: bpy.types.Image, src_chan: Channel):
+        self.image = image
+        self.src_chan = src_chan
+
+class FillWhite:
+    """Fills a channel with all ones (1.0)."""
+    pass
 
 
 class ExportImage:
-    """Custom image class that allows manipulation and encoding of images"""
-    # FUTURE_WORK: as a method to allow the node graph to be better supported, we could model some of
-    # the node graph elements with numpy functions
+    """Custom image class.
 
-    def __init__(self, img: typing.Union[np.ndarray, typing.List[np.ndarray]], max_channels: int = 4,\
-            blender_image: bpy.types.Image = None, has_alpha: bool = False):
-        if isinstance(img, list):
-            np.stack(img, axis=2)
+    An image is represented by giving a description of how to fill its red,
+    green, blue, and alpha channels. For example:
 
-        if len(img.shape) == 2:
-            # images must always have a channels dimension
-            img = np.expand_dims(img, axis=2)
+        self.fills = {
+            Channel.R: FillImage(image=bpy.data.images['Im1'], src_chan=Channel.B),
+            Channel.G: FillWhite(),
+        }
 
-        if not len(img.shape) == 3 or img.shape[2] > 4:
-            raise RuntimeError("Cannot construct an export image from an array of shape {}".format(img.shape))
+    This says that the ExportImage's R channel should be filled with the B
+    channel of the Blender image 'Im1', and the ExportImage's G channel
+    should be filled with all 1.0s. Undefined channels mean we don't care
+    what values that channel has.
 
-        self._img = img
-        self._max_channels = max_channels
-        self._blender_image = blender_image
-        self._has_alpha = has_alpha
+    This is flexible enough to handle the case where eg. the user used the R
+    channel of one image as the metallic value and the G channel of another
+    image as the roughness, and we need to synthesize an ExportImage that
+    packs those into the B and G channels for glTF.
 
-    def set_alpha(self, alpha: bool):
-        self._has_alpha = alpha
+    Storing this description (instead of raw pixels) lets us make more
+    intelligent decisions about how to encode the image.
+    """
 
-    @classmethod
-    def from_blender_image(cls, blender_image: bpy.types.Image):
-        img = np.array(blender_image.pixels[:])
-        img = img.reshape((blender_image.size[0], blender_image.size[1], blender_image.channels))
-        has_alpha = blender_image.depth == 32
-        return ExportImage(img=img, blender_image=blender_image, has_alpha=has_alpha)
+    def __init__(self):
+        self.fills = {}
 
-    @classmethod
-    def white_image(cls, width, height, num_channels: int = 4):
-        img = np.ones((width, height, num_channels))
-        return ExportImage(img=img)
+    @staticmethod
+    def from_blender_image(image: bpy.types.Image):
+        export_image = ExportImage()
+        for chan in range(image.channels):
+            export_image.fill_image(image, dst_chan=chan, src_chan=chan)
+        return export_image
 
-    def split_channels(self):
-        """return a list of numpy arrays where each list element corresponds to one image channel (r,g?,b?,a?)"""
-        return np.split(self._img, self._img.shape[2], axis=2)
+    def fill_image(self, image: bpy.types.Image, dst_chan: Channel, src_chan: Channel):
+        self.fills[dst_chan] = FillImage(image, src_chan)
 
-    @property
-    def img(self) -> np.ndarray:
-        return self._img
+    def fill_white(self, dst_chan: Channel):
+        self.fills[dst_chan] = FillWhite()
 
-    @property
-    def shape(self):
-        return self._img.shape
+    def is_filled(self, chan: Channel) -> bool:
+        return chan in self.fills
 
-    @property
-    def width(self):
-        return self.shape[0]
+    def __on_happy_path(self) -> bool:
+        # Whether there is an existing Blender image we can use for this
+        # ExportImage because all the channels come from the matching
+        # channel of that image, eg.
+        #
+        #     self.fills = {
+        #         Channel.R: FillImage(image=im, src_chan=Channel.R),
+        #         Channel.G: FillImage(image=im, src_chan=Channel.G),
+        #     }
+        return (
+            all(isinstance(fill, FillImage) for fill in self.fills.values()) and
+            all(dst_chan == fill.src_chan for dst_chan, fill in self.fills.items()) and
+            len(set(fill.image.name for fill in self.fills.values())) == 1
+        )
 
-    @property
-    def height(self):
-        return self.shape[1]
+    def encode(self, mime_type: Optional[str]) -> bytes:
+        self.mime_type = mime_type
 
-    @property
-    def channels(self):
-        return self.shape[2]
+        # All channels are white or undefined
+        # (I don't think this should happen...)
+        if all(isinstance(fill, FillWhite) for fill in self.fills.values()):
+            return self.__encode_blank()
 
-    def __getitem__(self, key):
-        """returns a new ExportImage with only the selected channels"""
-        return ExportImage(self._img[:, :, key])
+        # Happy path = we can just use an existing Blender image
+        if self.__on_happy_path():
+            return self.__encode_happy()
 
-    def __setitem__(self, key, value):
-        """set the selected channels to a new value"""
-        if isinstance(key, slice):
-            self._img[:, :, key] = value.img
-        else:
-            self._img[:, :, key] = value.img[:, :, 0]
+        # Unhappy path = we need to create the image self.fills describes.
+        return self.__encode_unhappy()
 
-    def append(self, other):
-        if self.channels + other.channels > self._max_channels:
-            raise RuntimeError("Cannot append image data to this image "
-                               "because the maximum number of channels is exceeded.")
+    def __encode_happy(self) -> bytes:
+        for fill in self.fills.values():
+            return self.__encode_from_image(fill.image)
 
-        self._img = np.concatenate([self.img, other.img], axis=2)
+    def __encode_unhappy(self) -> bytes:
+        # This will be a numpy array we fill in with pixel data.
+        result = None
 
-    def __add__(self, other):
-        self.append(other)
+        img_fills = {
+            chan: fill
+            for chan, fill in self.fills.items()
+            if isinstance(fill, FillImage)
+        }
+        # Loop over images instead of dst_chans; ensures we only decode each
+        # image once even if it's used in multiple channels.
+        image_names = list(set(fill.image.name for fill in img_fills.values()))
+        for image_name in image_names:
+            image = bpy.data.images[image_name]
 
-    def encode(self, mime_type: typing.Optional[str]) -> bytes:
+            if result is None:
+                result = np.ones((image.size[0], image.size[1], 4), np.float32)
+            # Images should all be the same size (should be guaranteed by
+            # gather_texture_info).
+            assert (image.size[0], image.size[1]) == result.shape[:2]
+
+            # Slow and eats all your memory.
+            pixels = np.array(image.pixels[:])
+
+            pixels = pixels.reshape((image.size[0], image.size[1], image.channels))
+
+            for dst_chan, img_fill in img_fills.items():
+                if img_fill.image == image:
+                    result[:, :, dst_chan] = pixels[:, :, img_fill.src_chan]
+
+            pixels = None  # GC this please
+
+        return self.__encode_from_numpy_array(result)
+
+    def __encode_blank(self) -> bytes:
+        # 1x1 white image
+        pixels = np.array([1.0, 1.0, 1.0, 1.0])
+        pixels = pixels.reshape((1, 1, 4))
+        return self.__encode_from_numpy_array(pixels)
+
+    def __encode_from_numpy_array(self, array: np.ndarray) -> bytes:
+        tmp_image = None
+        try:
+            tmp_image = bpy.data.images.new(
+                "##gltf-export:tmp-image##",
+                width=array.shape[0],
+                height=array.shape[1],
+                alpha=Channel.A in self.fills,
+            )
+            assert tmp_image.channels == 4  # 4 regardless of the alpha argument above.
+
+            # Also slow and eats all your memory.
+            tmp_image.pixels = array.flatten().tolist()
+
+            return self.__encode_from_image(tmp_image)
+
+        finally:
+            if tmp_image is not None:
+                bpy.data.images.remove(tmp_image, do_unlink=True)
+
+    def __encode_from_image(self, image: bpy.types.Image) -> bytes:
         file_format = {
             "image/jpeg": "JPEG",
             "image/png": "PNG"
-        }.get(mime_type, "PNG")
+        }.get(self.mime_type, "PNG")
 
-        if self._blender_image is not None and file_format == self._blender_image.file_format:
-            src_path = bpy.path.abspath(self._blender_image.filepath_raw)
+        # For images already on disk, skip saving them and just read them back.
+        # (What about if the image has changed on disk though?)
+        if file_format == image.file_format and not image.is_dirty:
+            src_path = bpy.path.abspath(image.filepath_raw)
             if os.path.isfile(src_path):
-                with open(src_path, "rb") as f:
-                    encoded_image = f.read()
-                return encoded_image
+                with open(src_path, 'rb') as f:
+                    return f.read()
 
-        image = bpy.data.images.new("TmpImage", width=self.width, height=self.height, alpha=self._has_alpha)
-        pixels = self._img.flatten().tolist()
-        image.pixels = pixels
-
-        # we just use blenders built in save mechanism, this can be considered slightly dodgy but currently is the only
-        # way to support
         with tempfile.TemporaryDirectory() as tmpdirname:
             tmpfilename = tmpdirname + "/img"
-            image.filepath_raw = tmpfilename
-            image.file_format = file_format
-            image.save()
+
+            # Save original values
+            orig_filepath_raw = image.filepath_raw
+            orig_file_format = image.file_format
+            try:
+                image.filepath_raw = tmpfilename
+                image.file_format = file_format
+
+                image.save()
+
+            finally:
+                # Restore original values
+                image.filepath_raw = orig_filepath_raw
+                image.file_format = orig_file_format
 
             with open(tmpfilename, "rb") as f:
-                encoded_image = f.read()
-
-        bpy.data.images.remove(image, do_unlink=True)
-
-        return encoded_image
+                return f.read()