This PR creates a "happy path" for exporting images where we don't have to round-trip through Python data, which is much faster and uses much less memory.

Fixes #790 and helps with export times generally.

## Problem

The glTF spec requires that eg. the metalllicRoughness texture have the metalness in B and the roughness in G. The exporter will assemble the appropriate image for this even if the material graph reads the metalness from, say, the R channel of one image and the roughness from the G channel of another.

To do this, *all* images are decoded to python data. The code sequence for this looks roughly like this:

    creating export image:
        pixel_data = np.array(blender_image.pixels[:])              (★1)

    pixel data may be modified to insert channels

    encoding export image:
        if created from blender image that's on disk:
            read back data from the disk

        tmp_image = bpy.data.images.new(...)
        tmp_image.pixels = pixel_data.img.flatten().tolist()        (★2)
        save tmp_image to disk and read it back

The lines marked with a ★ are slow and eat up loooots of memory. Eg. `blender_image.pixels[:]` is going to allocate a tuple of width×height×channels floats. It's really bad.

As you can see, sometimes it was possible to skip ★2 but everything did ★1.

However, for many images, like all color textures, normals, or metal/rough/occlusion textures that are already packed the way the exporter likes, this is totally wasted effort: the image data is converted to python data, nothing gets changed, and the python data is converted back to a new temp image, which is then encoded to disk and read back.

In this case it would have been better to just encode the original image!

## Solution

Change `ExportImage` to store some declarative data saying how it's channels should be filled in (ex. it says its R channel should be filled by the G channel of such-and-such Blender image) instead of storing a numpy array. I tried to write a good docstring here, so see it for details.

Then when it comes time to encode, we examine that data and determine whether all we need to do is save some existing Blender image (happy path), or whether we need to do the pixel manipulation to repack the channels (unhappy path).

The unhappy path is basically the same as before, but the happy path can be significantly better :)

## Results

Results seem pretty good. Obviously they depend on how much you can stay on the happy path but since that should include the most common texture types (base color, normal, etc.) it should be pretty often.

Export times:

| | Before | After|
|:---|---|---|
| #790 (8192x8192 normalmap) | OOM | 3s |
| DamagedHelmet | 15s | 7s |
| VC | 17s | 6s |
| Sponza | 91s | 38s |

Reviews and testing are appreciated!