@julienduroure 2019-11-25
I will have a look, but because material/texture is not my main domain of expertise, I hope @emackey or @donmccurdy can have a look