@scurest 2019-12-03
I wrote a [bash script](https://gist.github.com/scurest/c10a60d362a3b9fe52eef0d55eea9471) for testing if anyone else wants to try. Make sure you have the faster-export-image branch, cd in the glTF-Blender-IO folder, and run `test.sh FILE`. FILE will be loaded and then exported with both master and the faster-export-image branch, and the .pngs will be compared with imagemagick.

So far the only difference I have found is desirable IMO: this branch will sometimes figure out it can use an existing Blender image without change, where master would have changed it.

For example, if you import MetalRoughSpheres.glb, you get this image

![blender](https://user-images.githubusercontent.com/11024420/70076147-c704ad80-15c3-11ea-87b8-525973fd3327.png)

On master, if you export, the red channel will have been filled with ones

![out](https://user-images.githubusercontent.com/11024420/70076243-fadfd300-15c3-11ea-907c-3555cc9539e2.png)

whereas on this branch it exports exactly as it was.