@scurest 2019-12-03
Oh, here's a difference: when Blender textures are missing, it was possible to have an empty ExportImage which would export as a single pixel image. In master, the image is just thrown away in this case. I'll add a commit to fix it.