@scurest 2019-12-05
So I tested it with dirty images (ie. images that have been painted on and not saved; in the Image Editor there will be a `*` next to the Image menu if the current image is dirty) and I found another bug. After export, any image that goes down the happy path will have its dirty flag cleared since it was saved to a temp file.

This can be fixed by just doing `image.pixels[0] = image.pixels[0]` on images that were dirty to set their dirty flags again. This is can be pretty slow for some reason. Luckily we only pay this cost if an image was dirty. I added a commit to do this.

Btw, you might have noticed that the first commit also fixed a bug in master where it would use an image from disk even if the image was dirty.