@scurest 2019-12-06
Saving a copy of an image turns out to be surprisingly hard :(

Besides clearing the dirty flag, saving images would also turn generated images into from-files-on-disk images, so I've given up on using `image.save()` to save images.

Instead I'm now using `image.save_render()` which doesn't modify the image at all so it won't have any of those problems. It does use the render settings from a scene though, so this now creates a temp scene, sets all the settings we need, and then uses `save_render()`. See the commit message on 685e9ec for details.

(Doing it this way is slightly slower unfortunately; Sponza has gone to about 40s vs 38s before; the others I tested are about the same.) 

The reason the tests are failing on 2.79 is that the test expects an RGB image and somewhere between 2.79 and 2.8 Blender learned to do

    if you call img.save_render()
        if the render color mode is RGBA
            if img is an RGB image
                save the image as an RGB file instead of an RGBA file

which makes 2.8 pass, but since 2.79 doesn't know how to do this, the PNG files it saves when it saves a copy will be RGBA images instead of RGB. As far as I could tell, there is no field that exposes if the image data is RGB or RGBA on a Blender image, so I don't know how I could fix this. (Other than having an unnecessary A channel though, I don't think there's anything actually wrong with 2.79).

I'd really love any help writing `_save_copy_of_image`. Does anyone else have any ideas? Is there some way to ask what the correct way to save a copy of an image from Python is?