@julienduroure 2019-12-06
>  saving images would also turn generated images into from-files-on-disk images

You can get back to generative using bpy.data.images['Untitled'].source = 'GENERATED'

> clearing the dirty flag

This is dirty too, but we can change pixels[0] to any value and back to original one, it will get back the  dirty flag