@scurest 2019-12-06
> You can get back to generative using bpy.data.images['Untitled'].source = 'GENERATED'

You can't if the image was dirty though. When you set `image.source = 'GENERATED'` the image gets recreated from the generation parameters, which means any changes you painted on top of it would be lost.

> This is dirty too, but we can change pixels[0] to any value and back to original one, it will get back the dirty flag

Yeah, just `image.pixels[0] = image.pixels[0]` works (though like I said before, it is pretty slow; seems anything touching `pixels` is).

The other problem is I don't know if this is an exhaustive list of things that are messed up by `image.save()`.