@emackey 2019-12-06
@scurest Is `image.save_render()` only blocked by the 2.79 behavior described in your [comment](https://github.com/KhronosGroup/glTF-Blender-IO/pull/798#issuecomment-562442020) above?  We'll be dropping 2.79 soon, as you know.

The other thing I thought of is that the UI actually has something similar to what we want, saving an image as a copy from a menu item.  There's a matching Python operator:

> bpy.ops.image.save_as(save_as_render=False, copy=True, filepath="/path/file.png", relative_path=True, show_multiview=False, use_multiview=False)

I know that addons are discouraged from using `ops` for stuff like this, but it means there's code in there somewhere ([here](https://developer.blender.org/diffusion/B/browse/master/source/blender/editors/space_image/image_ops.c$1919)? [here](https://developer.blender.org/diffusion/B/browse/master/source/blender/blenkernel/intern/image_save.c$161)?) for this, even if there doesn't seem to be a non-ops Python binding for it.

Let's not let version 2.79 hold us back if that's the only stumbling block on `save_render` though.