@scurest 2019-12-06
>  There's a matching Python operator

Yeah, I'd love to use it but I can't figure out how. Do you know?

> Is image.save_render() only blocked by the 2.79 behavior described in your comment above?

AFAIK yes.

I'm not sure if the 2.79 behaviour should really be considered blocking. It's unfortunate that an alpha channel is added when it isn't needed, but it doesn't really hurt anything except the file size. This also only happens in what should be the uncommon case, when we can't find an existing file to use for the image (eg. it was packed from channels, it was dirty, it was generated).

If another way can't be found... If the test could be relaxed to only check if the images are pixel-for-pixel equal instead of checking if the files are byte-for-byte equal, it should pass on 2.79 too. (I think that would be a good idea anyway; checking if the PNGs files are byte-for-byte equal is too restrictive. For example, the tests will break even if all we do is change the compression setting for PNGs since the file will get encoded differently even though obviously the pixels don't change.)