@julienduroure 2019-12-06
> Yeah, I'd love to use it but I can't figure out how. Do you know?

You have to override the context, see https://docs.blender.org/api/current/bpy.ops.html#overriding-context