@emackey 2019-12-06
> > There's a matching Python operator
> 
> Yeah, I'd love to use it but I can't figure out how. Do you know?

I haven't used these much but my understanding is that `bpy.ops.*` is tied specifically to user UI operations, such as saving an image from an Image Editor or UV Editor UI panel.  This may be a dead-end.

> checking if the PNGs files are byte-for-byte equal is too restrictive

Yeah, I put in the byte-for-byte check when I was looking to test a "happy path" where the file was simply copied from the source, and then I probably reused it and never wrote a more detailed pixel-for-pixel check.  Some sort of relaxing of tests would be fine with me, we could even isolate this test to only Blender 2.8x and above, and skip it for 2.79.